feat: initial scaffolding of webpack rule
This is a strange way to do it, where we just depend on rules_nodejs core package.
There is no npm_install or nodejs_binary rule there yet,
and no linker, so we just depend on locally running npm install first.
This would have to be distributed as a bazel rule with the webpack dependency totally
hidden as a toolchain by vendoring its code.

Probably not the right shape